+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `timeout` input.
+
 ## [v1.0.0](https://github.com/allenai/beaker-run-action/releases/tag/v1.0.0) - 2022-04-10
 
 ### Added